master

+++ projects/include/protocol_consts.hpp
@@ -36,7 +36,7 @@ namespace protocol
 
     static const uint8_t REPLY_WRITE_METADATA_SIZE = REPLY_READ_METADATA_SIZE;
 
-    static const uint8_t REQUEST_METADATA_SIZE = sizeof(uint8_t) + 2 * sizeof(uint64_t);
+    static const uint64_t REQUEST_METADATA_SIZE = sizeof(uint8_t) + 2 * sizeof(uint64_t);
     
     static const uint64_t REPLY_READ_SIZE = BLOCK_SIZE + REPLY_READ_METADATA_SIZE;
     
@@ -44,12 +44,12 @@ namespace protocol
 
     static const uint64_t RECV_BLOCK_SIZE = BLOCK_SIZE + REQUEST_METADATA_SIZE;
 
-    static const uint8_t OPERATION_TYPE_OFFSET = 0;
-    static const uint8_t REQUEST_ID_OFFSET = 1;
-    static const uint8_t BLOCK_ID_OFFSET = 9;
-    static const uint8_t ERROR_CODE_OFFSET = 9;
-    static const uint8_t READ_DATA_BLOCK_OFFSET = 10;
-    static const uint8_t WRITE_DATA_BLOCK_OFFSET = 17;
+    static const uint64_t OPERATION_TYPE_OFFSET = 0;
+    static const uint64_t REQUEST_ID_OFFSET = 1;
+    static const uint64_t BLOCK_ID_OFFSET = 9;
+    static const uint64_t ERROR_CODE_OFFSET = 9;
+    static const uint64_t READ_DATA_BLOCK_OFFSET = 10;
+    static const uint64_t WRITE_DATA_BLOCK_OFFSET = 17;
 
 } // namespace protocol
